Add `model_adaptation` to `observation_type` implementation
