+++ avalanche/training/supervised/cumulative.py
@@ -7,7 +7,7 @@ from torch.utils.data import ConcatDataset
 from avalanche.benchmarks.utils import AvalancheConcatDataset
 from avalanche.training.plugins.evaluation import default_evaluator
 from avalanche.training.plugins import SupervisedPlugin, EvaluationPlugin
-from avalanche.training.templates.supervised import SupervisedTemplate
+from avalanche.training.templates import SupervisedTemplate
 
 
 class Cumulative(SupervisedTemplate):
