+++ avalanche/training/supervised/deep_slda.py
@@ -5,7 +5,7 @@ import os
 import torch
 
 from avalanche.training.plugins import SupervisedPlugin
-from avalanche.training.templates.supervised import SupervisedTemplate
+from avalanche.training.templates import SupervisedTemplate
 from avalanche.training.plugins.evaluation import default_evaluator
 from avalanche.models.dynamic_modules import MultiTaskModule
 from avalanche.models import FeatureExtractorBackbone
