+++ avalanche/training/supervised/joint_training.py
@@ -17,7 +17,7 @@ from torch.optim import Optimizer
 from avalanche.benchmarks.scenarios import ClassificationExperience
 from avalanche.benchmarks.utils import AvalancheConcatDataset
 from avalanche.training.plugins.evaluation import default_evaluator
-from avalanche.training.templates.supervised import SupervisedTemplate
+from avalanche.training.templates import SupervisedTemplate
 from avalanche.models import DynamicModule
 
 if TYPE_CHECKING:
