+++ examples/lamaml_cifar100.py
@@ -31,7 +31,8 @@ def main(args):
     )
 
     # --- SCENARIO CREATION
-    scenario = SplitCIFAR100(n_experiences=20, return_task_id=True)
+    scenario = SplitCIFAR100(n_experiences=20, return_task_id=True,
+                             class_ids_from_zero_in_each_exp=True)
     config = {"scenario": "SplitCIFAR100"}
 
     # MODEL CREATION
